init_chroma/requirements-vdb.txt
icecream==2.1.3
langchain-chroma==0.1.1
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.10
pandas==2.2.2
tqdm==4.66.4
python-dotenv==1.0.1
openpyxl==3.1.2